refactor: ♻️ #9 汎用的なクラスの保存場所を変更する
