+++ games/common/constants/game.ts
@@ -1,0 +1,10 @@
+const game = {
+  table: {
+    betDenominations: [5, 20, 50, 100]
+  },
+  player: {
+    chips: 1000
+  }
+} as const;
+
+export default game;
